restructure nixui folder
