+++ nixui/graphics/icon.py
@@ -1,4 +1,5 @@
 from PyQt5 import QtGui
+
 import pkgutil
